requirements.txt
fastapi
uvicorn
faiss-cpu
pypdf
youtube-transcript-api
python-dotenv
langchain
langchain_core
langchain_community
langchain_groq
pydantic
langchain_huggingface
bs4